Move interaction with point charges logic to EMLEPC class
