+++ emle/models/_emle.py
@@ -39,6 +39,8 @@ from typing import Optional, Tuple, List
 from . import _patches
 from . import EMLEBase as _EMLEBase
 
+from . import EMLEPC as _EMLEPC
+
 # Monkey-patch the TorchANI BuiltInModel and BuiltinEnsemble classes so that
 # they call self.aev_computer using args only to allow forward hooks to work
 # with TorchScript.
@@ -450,177 +452,23 @@ class EMLE(_torch.nn.Module):
                 q_core, dtype=charges_mm.dtype, device=self._device
             )
 
-        mesh_data = self._get_mesh_data(xyz_qm_bohr, xyz_mm_bohr, s)
+        mesh_data = _EMLEPC._get_mesh_data(xyz_qm_bohr, xyz_mm_bohr, s)
         if self._method == "mechanical":
             q_core = q_core + q_val
             q_val = _torch.zeros_like(
                 q_core, dtype=charges_mm.dtype, device=self._device
             )
-        vpot_q_core = self._get_vpot_q(q_core, mesh_data[0])
-        vpot_q_val = self._get_vpot_q(q_val, mesh_data[1])
+        vpot_q_core = _EMLEPC._get_vpot_q(q_core, mesh_data[0])
+        vpot_q_val = _EMLEPC._get_vpot_q(q_val, mesh_data[1])
         vpot_static = vpot_q_core + vpot_q_val
         E_static = _torch.sum(vpot_static @ charges_mm)
 
         # Compute the induced energy.
         if self._method == "electrostatic":
-            mu_ind = self._get_mu_ind(A_thole, mesh_data, charges_mm, s)
-            vpot_ind = self._get_vpot_mu(mu_ind, mesh_data[2])
+            mu_ind = _EMLEPC._get_mu_ind(A_thole, mesh_data, charges_mm, s)
+            vpot_ind = _EMLEPC._get_vpot_mu(mu_ind, mesh_data[2])
             E_ind = _torch.sum(vpot_ind @ charges_mm) * 0.5
         else:
             E_ind = _torch.tensor(0.0, dtype=charges_mm.dtype, device=self._device)
 
         return _torch.stack([E_static, E_ind])
-
-    def _get_mu_ind(
-        self,
-        A,
-        mesh_data: Tuple[Tensor, Tensor, Tensor],
-        q,
-        s,
-    ):
-        """
-        Internal method, calculates induced atomic dipoles
-        (Eq. 20 in 10.1021/acs.jctc.2c00914)
-
-        Parameters
-        ----------
-
-        A: torch.Tensor (N_ATOMS * 3, N_ATOMS * 3)
-            The A matrix for induced dipoles prediction.
-
-        mesh_data: mesh_data object (output of self._get_mesh_data)
-
-        q: torch.Tensor (N_MM_ATOMS,)
-            MM point charges.
-
-        s: torch.Tensor (N_QM_ATOMS,)
-            MBIS valence shell widths.
-
-        q_val: torch.Tensor (N_QM_ATOMS,)
-            MBIS valence charges.
-
-        Returns
-        -------
-
-        result: torch.Tensor (N_ATOMS, 3)
-            Array of induced dipoles
-        """
-
-        r = 1.0 / mesh_data[0]
-        f1 = self._get_f1_slater(r, s[:, None] * 2.0)
-        fields = _torch.sum(mesh_data[2] * f1[:, :, None] * q[:, None], dim=1).flatten()
-
-        mu_ind = _torch.linalg.solve(A, fields)
-        return mu_ind.reshape((-1, 3))
-
-    @staticmethod
-    def _get_vpot_q(q, T0):
-        """
-        Internal method to calculate the electrostatic potential.
-
-        Parameters
-        ----------
-
-        q: torch.Tensor (N_MM_ATOMS,)
-            MM point charges.
-
-        T0: torch.Tensor (N_QM_ATOMS, max_mm_atoms)
-            T0 tensor for QM atoms over MM atom positions.
-
-        Returns
-        -------
-
-        result: torch.Tensor (max_mm_atoms)
-            Electrostatic potential over MM atoms.
-        """
-        return _torch.sum(T0 * q[:, None], dim=0)
-
-    @staticmethod
-    def _get_vpot_mu(mu, T1):
-        """
-        Internal method to calculate the electrostatic potential generated
-        by atomic dipoles.
-
-        Parameters
-        ----------
-
-        mu: torch.Tensor (N_ATOMS, 3)
-            Atomic dipoles.
-
-        T1: torch.Tensor (N_ATOMS, max_mm_atoms, 3)
-            T1 tensor for QM atoms over MM atom positions.
-
-        Returns
-        -------
-
-        result: torch.Tensor (max_mm_atoms)
-            Electrostatic potential over MM atoms.
-        """
-        return -_torch.tensordot(T1, mu, ((0, 2), (0, 1)))
-
-    @classmethod
-    def _get_mesh_data(cls, xyz, xyz_mesh, s):
-        """
-        Internal method, calculates mesh_data object.
-
-        Parameters
-        ----------
-
-        xyz: torch.Tensor (N_ATOMS, 3)
-            Atomic positions.
-
-        xyz_mesh: torch.Tensor (max_mm_atoms, 3)
-            MM positions.
-
-        s: torch.Tensor (N_ATOMS,)
-            MBIS valence widths.
-        """
-        rr = xyz_mesh[None, :, :] - xyz[:, None, :]
-        r = _torch.linalg.norm(rr, ord=2, dim=2)
-
-        return (1.0 / r, cls._get_T0_slater(r, s[:, None]), -rr / r[:, :, None] ** 3)
-
-    @classmethod
-    def _get_f1_slater(cls, r, s):
-        """
-        Internal method, calculates damping factors for Slater densities.
-
-        Parameters
-        ----------
-
-        r: torch.Tensor (N_ATOMS, max_mm_atoms)
-            Distances from QM to MM atoms.
-
-        s: torch.Tensor (N_ATOMS,)
-            MBIS valence widths.
-
-        Returns
-        -------
-
-        result: torch.Tensor (N_ATOMS, max_mm_atoms)
-        """
-        return (
-            cls._get_T0_slater(r, s) * r
-            - _torch.exp(-r / s) / s * (0.5 + r / (s * 2)) * r
-        )
-
-    @staticmethod
-    def _get_T0_slater(r, s):
-        """
-        Internal method, calculates T0 tensor for Slater densities.
-
-        Parameters
-        ----------
-
-        r: torch.Tensor (N_ATOMS, max_mm_atoms)
-            Distances from QM to MM atoms.
-
-        s: torch.Tensor (N_ATOMS,)
-            MBIS valence widths.
-
-        Returns
-        -------
-
-        results: torch.Tensor (N_ATOMS, max_mm_atoms)
-        """
-        return (1 - (1 + r / (s * 2)) * _torch.exp(-r / s)) / r
